// src/richtexteditor/src/common/interface.ts
export interface IToolbarItemModel extends ItemModel {
    command?: string
    subCommand?: string
}
// ...
export interface IToolsItems {
    id: string
    icon?: string
    tooltip?: string
    command?: string
    subCommand?: string
    value?: string
}
// ...
export interface IToolbarItems {
    template?: string
    tooltipText?: string
    command?: string
    subCommand?: string
    undo?: boolean
    click?: EmitType<ClickEventArgs>
}